chore(calendar): Apply MS reviews
