+++ src/Uno.UI/DirectUI/ITreeBuilder.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Linq;
+
+namespace DirectUI
+{
+	internal interface ITreeBuilder
+	{
+		public bool IsBuildTreeSuspended { get; }
+
+		public bool IsRegisteredForCallbacks { get; set; }
+
+		public bool BuildTree();
+	}
+}
